+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/_container_service_client.py
@@ -9,17 +9,19 @@
 # regenerated.
 # --------------------------------------------------------------------------
 
-from typing import Any, Optional, TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING, cast
 from typing_extensions import Self
 
 from azure.core.pipeline import policies
+from azure.core.settings import settings
 from azure.mgmt.core import ARMPipelineClient
 from azure.mgmt.core.policies import ARMAutoResourceProviderRegistrationPolicy
+from azure.mgmt.core.tools import get_arm_endpoints
 from azure.profiles import KnownProfiles, ProfileDefinition
 from azure.profiles.multiapiclient import MultiApiClientMixin
 
 from ._configuration import ContainerServiceClientConfiguration
-from ._serialization import Deserializer, Serializer
+from ._utils.serialization import Deserializer, Serializer
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
@@ -56,7 +58,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
-    DEFAULT_API_VERSION = '2025-01-01'
+    DEFAULT_API_VERSION = '2025-02-01'
     _PROFILE_TAG = "azure.mgmt.containerservice.ContainerServiceClient"
     LATEST_PROFILE = ProfileDefinition({
         _PROFILE_TAG: {
@@ -72,13 +74,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         credential: "TokenCredential",
         subscription_id: str,
         api_version: Optional[str]=None,
-        base_url: str = "https://management.azure.com",
+        base_url: Optional[str] = None,
         profile: KnownProfiles=KnownProfiles.default,
         **kwargs: Any
     ):
         if api_version:
             kwargs.setdefault('api_version', api_version)
-        self._config = ContainerServiceClientConfiguration(credential, subscription_id, **kwargs)
+        _cloud = kwargs.pop("cloud_setting", None) or settings.current.azure_cloud  # type: ignore
+        _endpoints = get_arm_endpoints(_cloud)
+        if not base_url:
+            base_url = _endpoints["resource_manager"]
+        credential_scopes = kwargs.pop("credential_scopes", _endpoints["credential_scopes"])
+        self._config = ContainerServiceClientConfiguration(credential, subscription_id, credential_scopes=credential_scopes, **kwargs)
         _policies = kwargs.pop("policies", None)
         if _policies is None:
             _policies = [
@@ -97,7 +104,7 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
                 policies.SensitiveHeaderCleanupPolicy(**kwargs) if self._config.redirect_policy else None,
                 self._config.http_logging_policy,
             ]
-        self._client: ARMPipelineClient = ARMPipelineClient(base_url=base_url, policies=_policies, **kwargs)
+        self._client: ARMPipelineClient = ARMPipelineClient(base_url=cast(str, base_url), policies=_policies, **kwargs)
         super(ContainerServiceClient, self).__init__(
             api_version=api_version,
             profile=profile
@@ -115,8 +122,8 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
            * 2019-09-30-preview: :mod:`v2019_09_30_preview.models<azure.mgmt.containerservice.v2019_09_30_preview.models>`
            * 2020-03-01: :mod:`v2020_03_01.models<azure.mgmt.containerservice.v2020_03_01.models>`
            * 2020-11-01: :mod:`v2020_11_01.models<azure.mgmt.containerservice.v2020_11_01.models>`
-           * 2025-01-01: :mod:`v2025_01_01.models<azure.mgmt.containerservice.v2025_01_01.models>`
-           * 2025-01-02-preview: :mod:`v2025_01_02_preview.models<azure.mgmt.containerservice.v2025_01_02_preview.models>`
+           * 2025-02-01: :mod:`v2025_02_01.models<azure.mgmt.containerservice.v2025_02_01.models>`
+           * 2025-02-02-preview: :mod:`v2025_02_02_preview.models<azure.mgmt.containerservice.v2025_02_02_preview.models>`
         """
         if api_version == '2017-07-01':
             from .v2017_07_01 import models
@@ -130,11 +137,11 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         elif api_version == '2020-11-01':
             from .v2020_11_01 import models
             return models
-        elif api_version == '2025-01-01':
-            from .v2025_01_01 import models
+        elif api_version == '2025-02-01':
+            from .v2025_02_01 import models
             return models
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview import models
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview import models
             return models
         raise ValueError("API version {} is not available".format(api_version))
 
@@ -144,18 +151,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2020-03-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2020_03_01.operations.AgentPoolsOperations>`
            * 2020-11-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2020_11_01.operations.AgentPoolsOperations>`
-           * 2025-01-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2025_01_01.operations.AgentPoolsOperations>`
-           * 2025-01-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.AgentPoolsOperations>`
+           * 2025-02-01: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2025_02_01.operations.AgentPoolsOperations>`
+           * 2025-02-02-preview: :class:`AgentPoolsOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.AgentPoolsOperations>`
         """
         api_version = self._get_api_version('agent_pools')
         if api_version == '2020-03-01':
             from .v2020_03_01.operations import AgentPoolsOperations as OperationClass
         elif api_version == '2020-11-01':
             from .v2020_11_01.operations import AgentPoolsOperations as OperationClass
-        elif api_version == '2025-01-01':
-            from .v2025_01_01.operations import AgentPoolsOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2025-02-01':
+            from .v2025_02_01.operations import AgentPoolsOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import AgentPoolsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'agent_pools'".format(api_version))
         self._config.api_version = api_version
@@ -165,11 +172,11 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def container_service(self):
         """Instance depends on the API version:
 
-           * 2025-01-02-preview: :class:`ContainerServiceOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.ContainerServiceOperations>`
+           * 2025-02-02-preview: :class:`ContainerServiceOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.ContainerServiceOperations>`
         """
         api_version = self._get_api_version('container_service')
-        if api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import ContainerServiceOperations as OperationClass
+        if api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import ContainerServiceOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'container_service'".format(api_version))
         self._config.api_version = api_version
@@ -193,11 +200,11 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def load_balancers(self):
         """Instance depends on the API version:
 
-           * 2025-01-02-preview: :class:`LoadBalancersOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.LoadBalancersOperations>`
+           * 2025-02-02-preview: :class:`LoadBalancersOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.LoadBalancersOperations>`
         """
         api_version = self._get_api_version('load_balancers')
-        if api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import LoadBalancersOperations as OperationClass
+        if api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import LoadBalancersOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'load_balancers'".format(api_version))
         self._config.api_version = api_version
@@ -207,14 +214,14 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def machines(self):
         """Instance depends on the API version:
 
-           * 2025-01-01: :class:`MachinesOperations<azure.mgmt.containerservice.v2025_01_01.operations.MachinesOperations>`
-           * 2025-01-02-preview: :class:`MachinesOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.MachinesOperations>`
+           * 2025-02-01: :class:`MachinesOperations<azure.mgmt.containerservice.v2025_02_01.operations.MachinesOperations>`
+           * 2025-02-02-preview: :class:`MachinesOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.MachinesOperations>`
         """
         api_version = self._get_api_version('machines')
-        if api_version == '2025-01-01':
-            from .v2025_01_01.operations import MachinesOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import MachinesOperations as OperationClass
+        if api_version == '2025-02-01':
+            from .v2025_02_01.operations import MachinesOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import MachinesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'machines'".format(api_version))
         self._config.api_version = api_version
@@ -224,14 +231,14 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def maintenance_configurations(self):
         """Instance depends on the API version:
 
-           * 2025-01-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2025_01_01.operations.MaintenanceConfigurationsOperations>`
-           * 2025-01-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.MaintenanceConfigurationsOperations>`
+           * 2025-02-01: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2025_02_01.operations.MaintenanceConfigurationsOperations>`
+           * 2025-02-02-preview: :class:`MaintenanceConfigurationsOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.MaintenanceConfigurationsOperations>`
         """
         api_version = self._get_api_version('maintenance_configurations')
-        if api_version == '2025-01-01':
-            from .v2025_01_01.operations import MaintenanceConfigurationsOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
+        if api_version == '2025-02-01':
+            from .v2025_02_01.operations import MaintenanceConfigurationsOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import MaintenanceConfigurationsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'maintenance_configurations'".format(api_version))
         self._config.api_version = api_version
@@ -241,11 +248,11 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def managed_cluster_snapshots(self):
         """Instance depends on the API version:
 
-           * 2025-01-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.ManagedClusterSnapshotsOperations>`
+           * 2025-02-02-preview: :class:`ManagedClusterSnapshotsOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.ManagedClusterSnapshotsOperations>`
         """
         api_version = self._get_api_version('managed_cluster_snapshots')
-        if api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
+        if api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import ManagedClusterSnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_cluster_snapshots'".format(api_version))
         self._config.api_version = api_version
@@ -257,20 +264,34 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2020-03-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2020_03_01.operations.ManagedClustersOperations>`
            * 2020-11-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2020_11_01.operations.ManagedClustersOperations>`
-           * 2025-01-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2025_01_01.operations.ManagedClustersOperations>`
-           * 2025-01-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.ManagedClustersOperations>`
+           * 2025-02-01: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2025_02_01.operations.ManagedClustersOperations>`
+           * 2025-02-02-preview: :class:`ManagedClustersOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.ManagedClustersOperations>`
         """
         api_version = self._get_api_version('managed_clusters')
         if api_version == '2020-03-01':
             from .v2020_03_01.operations import ManagedClustersOperations as OperationClass
         elif api_version == '2020-11-01':
             from .v2020_11_01.operations import ManagedClustersOperations as OperationClass
-        elif api_version == '2025-01-01':
-            from .v2025_01_01.operations import ManagedClustersOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2025-02-01':
+            from .v2025_02_01.operations import ManagedClustersOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import ManagedClustersOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'managed_clusters'".format(api_version))
+        self._config.api_version = api_version
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
+
+    @property
+    def namespaces(self):
+        """Instance depends on the API version:
+
+           * 2025-02-02-preview: :class:`NamespacesOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.NamespacesOperations>`
+        """
+        api_version = self._get_api_version('namespaces')
+        if api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import NamespacesOperations as OperationClass
+        else:
+            raise ValueError("API version {} does not have operation group 'namespaces'".format(api_version))
         self._config.api_version = api_version
         return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)), api_version)
 
@@ -292,11 +313,11 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def operation_status_result(self):
         """Instance depends on the API version:
 
-           * 2025-01-02-preview: :class:`OperationStatusResultOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.OperationStatusResultOperations>`
+           * 2025-02-02-preview: :class:`OperationStatusResultOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.OperationStatusResultOperations>`
         """
         api_version = self._get_api_version('operation_status_result')
-        if api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import OperationStatusResultOperations as OperationClass
+        if api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import OperationStatusResultOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'operation_status_result'".format(api_version))
         self._config.api_version = api_version
@@ -308,18 +329,18 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
 
            * 2020-03-01: :class:`Operations<azure.mgmt.containerservice.v2020_03_01.operations.Operations>`
            * 2020-11-01: :class:`Operations<azure.mgmt.containerservice.v2020_11_01.operations.Operations>`
-           * 2025-01-01: :class:`Operations<azure.mgmt.containerservice.v2025_01_01.operations.Operations>`
-           * 2025-01-02-preview: :class:`Operations<azure.mgmt.containerservice.v2025_01_02_preview.operations.Operations>`
+           * 2025-02-01: :class:`Operations<azure.mgmt.containerservice.v2025_02_01.operations.Operations>`
+           * 2025-02-02-preview: :class:`Operations<azure.mgmt.containerservice.v2025_02_02_preview.operations.Operations>`
         """
         api_version = self._get_api_version('operations')
         if api_version == '2020-03-01':
             from .v2020_03_01.operations import Operations as OperationClass
         elif api_version == '2020-11-01':
             from .v2020_11_01.operations import Operations as OperationClass
-        elif api_version == '2025-01-01':
-            from .v2025_01_01.operations import Operations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import Operations as OperationClass
+        elif api_version == '2025-02-01':
+            from .v2025_02_01.operations import Operations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import Operations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'operations'".format(api_version))
         self._config.api_version = api_version
@@ -330,16 +351,16 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2020-11-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2020_11_01.operations.PrivateEndpointConnectionsOperations>`
-           * 2025-01-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2025_01_01.operations.PrivateEndpointConnectionsOperations>`
-           * 2025-01-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.PrivateEndpointConnectionsOperations>`
+           * 2025-02-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2025_02_01.operations.PrivateEndpointConnectionsOperations>`
+           * 2025-02-02-preview: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.PrivateEndpointConnectionsOperations>`
         """
         api_version = self._get_api_version('private_endpoint_connections')
         if api_version == '2020-11-01':
             from .v2020_11_01.operations import PrivateEndpointConnectionsOperations as OperationClass
-        elif api_version == '2025-01-01':
-            from .v2025_01_01.operations import PrivateEndpointConnectionsOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2025-02-01':
+            from .v2025_02_01.operations import PrivateEndpointConnectionsOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import PrivateEndpointConnectionsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_endpoint_connections'".format(api_version))
         self._config.api_version = api_version
@@ -350,16 +371,16 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2020-11-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2020_11_01.operations.PrivateLinkResourcesOperations>`
-           * 2025-01-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2025_01_01.operations.PrivateLinkResourcesOperations>`
-           * 2025-01-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.PrivateLinkResourcesOperations>`
+           * 2025-02-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2025_02_01.operations.PrivateLinkResourcesOperations>`
+           * 2025-02-02-preview: :class:`PrivateLinkResourcesOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.PrivateLinkResourcesOperations>`
         """
         api_version = self._get_api_version('private_link_resources')
         if api_version == '2020-11-01':
             from .v2020_11_01.operations import PrivateLinkResourcesOperations as OperationClass
-        elif api_version == '2025-01-01':
-            from .v2025_01_01.operations import PrivateLinkResourcesOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2025-02-01':
+            from .v2025_02_01.operations import PrivateLinkResourcesOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import PrivateLinkResourcesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'private_link_resources'".format(api_version))
         self._config.api_version = api_version
@@ -370,16 +391,16 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
         """Instance depends on the API version:
 
            * 2020-11-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2020_11_01.operations.ResolvePrivateLinkServiceIdOperations>`
-           * 2025-01-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2025_01_01.operations.ResolvePrivateLinkServiceIdOperations>`
-           * 2025-01-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2025-02-01: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2025_02_01.operations.ResolvePrivateLinkServiceIdOperations>`
+           * 2025-02-02-preview: :class:`ResolvePrivateLinkServiceIdOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.ResolvePrivateLinkServiceIdOperations>`
         """
         api_version = self._get_api_version('resolve_private_link_service_id')
         if api_version == '2020-11-01':
             from .v2020_11_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
-        elif api_version == '2025-01-01':
-            from .v2025_01_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2025-02-01':
+            from .v2025_02_01.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import ResolvePrivateLinkServiceIdOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'resolve_private_link_service_id'".format(api_version))
         self._config.api_version = api_version
@@ -389,14 +410,14 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def snapshots(self):
         """Instance depends on the API version:
 
-           * 2025-01-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2025_01_01.operations.SnapshotsOperations>`
-           * 2025-01-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.SnapshotsOperations>`
+           * 2025-02-01: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2025_02_01.operations.SnapshotsOperations>`
+           * 2025-02-02-preview: :class:`SnapshotsOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.SnapshotsOperations>`
         """
         api_version = self._get_api_version('snapshots')
-        if api_version == '2025-01-01':
-            from .v2025_01_01.operations import SnapshotsOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import SnapshotsOperations as OperationClass
+        if api_version == '2025-02-01':
+            from .v2025_02_01.operations import SnapshotsOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import SnapshotsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'snapshots'".format(api_version))
         self._config.api_version = api_version
@@ -406,14 +427,14 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def trusted_access_role_bindings(self):
         """Instance depends on the API version:
 
-           * 2025-01-01: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2025_01_01.operations.TrustedAccessRoleBindingsOperations>`
-           * 2025-01-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.TrustedAccessRoleBindingsOperations>`
+           * 2025-02-01: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2025_02_01.operations.TrustedAccessRoleBindingsOperations>`
+           * 2025-02-02-preview: :class:`TrustedAccessRoleBindingsOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.TrustedAccessRoleBindingsOperations>`
         """
         api_version = self._get_api_version('trusted_access_role_bindings')
-        if api_version == '2025-01-01':
-            from .v2025_01_01.operations import TrustedAccessRoleBindingsOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
+        if api_version == '2025-02-01':
+            from .v2025_02_01.operations import TrustedAccessRoleBindingsOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import TrustedAccessRoleBindingsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_role_bindings'".format(api_version))
         self._config.api_version = api_version
@@ -423,14 +444,14 @@ class ContainerServiceClient(MultiApiClientMixin, _SDKClient):
     def trusted_access_roles(self):
         """Instance depends on the API version:
 
-           * 2025-01-01: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2025_01_01.operations.TrustedAccessRolesOperations>`
-           * 2025-01-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2025_01_02_preview.operations.TrustedAccessRolesOperations>`
+           * 2025-02-01: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2025_02_01.operations.TrustedAccessRolesOperations>`
+           * 2025-02-02-preview: :class:`TrustedAccessRolesOperations<azure.mgmt.containerservice.v2025_02_02_preview.operations.TrustedAccessRolesOperations>`
         """
         api_version = self._get_api_version('trusted_access_roles')
-        if api_version == '2025-01-01':
-            from .v2025_01_01.operations import TrustedAccessRolesOperations as OperationClass
-        elif api_version == '2025-01-02-preview':
-            from .v2025_01_02_preview.operations import TrustedAccessRolesOperations as OperationClass
+        if api_version == '2025-02-01':
+            from .v2025_02_01.operations import TrustedAccessRolesOperations as OperationClass
+        elif api_version == '2025-02-02-preview':
+            from .v2025_02_02_preview.operations import TrustedAccessRolesOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'trusted_access_roles'".format(api_version))
         self._config.api_version = api_version