+++ src/aks-preview/HISTORY.rst
@@ -12,18 +12,26 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+14.0.0b5
+++++++++
+* Add `az aks create/update --enable-retina-flow-logs` and `az aks update --disable-retina-flow-logs` commands.
+
+14.0.0b4
+++++++++
+* Vendor new SDK and bump API version to 2025-02-02-preview.
+
 14.0.0b3
 +++++++
-* Add `az aks create/update --enable-retina-flow-logs` and `az aks update --disable-retina-flow-logs` commands.
+* Support `az aks loadbalancer show/list/add/update/delete/rebalance` commands.
 
 14.0.0b2
 +++++++
 * Update the `disable-egress-gateway` subcommand to fix `--help` output for the `az aks mesh` command.
-* Vendor new SDK and bump API version to 2025-01-02-preview.
 * Modify behavior for `--nodepool-initialization-taints` to ignore taints with hard effects on node pools with system mode when creating or updating a cluster.
 
 14.0.0b1
 +++++++
+* Vendor new SDK and bump API version to 2025-01-02-preview.
 * [BREAKING CHANGE] Rename `--enable-addon-autoscaling` to `--enable-optimized-addon-scaling` to `az aks create` commands.
 * [BREAKING CHANGE] Rename `--enable-addon-autoscaling` to `--enable-optimized-addon-scaling` and `--disable-addon-autoscaling` to `--disable-optimized-addon-scaling` to `az aks update` commands.